Add hook to manage window dimensions
